set maximum SPI speed
The raspberry linux kernel updated the default spi frequency to 125MHz:

kernel: BCM270X_DT: Set spidev spi-max-frequency to 125MHz.
See raspberrypi/linux#2165.

The WS2801 LEDs can only handle 10 MHz, but when not set the default value is used, which is too much.
This commit explicit includes the spi package and sets the max speed for the spidev